requirements.txt
python-telegram-bot
supabase-py
python-dotenv